requirements.txt
decorator==4.3.0
Django==1.11
django-common-helpers==0.9.1
django-cors-headers==2.1.0
django-enumfield==1.3b2
django-filter==1.0.4
django-invitations==1.9.1
django-phonenumber-field==1.3.0
django-unused-media==0.1.10
djangorestframework==3.6.4
django-heroku==0.3.1
djutils==0.3.2
dnspython==1.15.0
docutils==0.14
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
imagesize==1.0.0
inflect==0.3.1
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
isodate==0.6.0
jsonfield==2.0.2
jsonpickle==0.9.5
jsonschema==2.6.0
oauth2client==3.0.0
oauthlib==2.1.0
olefile==0.44
openpyxl==2.5.3
packaging==17.1
Pillow==4.2.1
psycopg2==2.7.3.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycryptodome==3.4.3
pyflakes==2.0.0
Pygments==2.2.0
PyJWT==1.5.3
pylint==2.1.1
pyparsing==2.2.0
pyproj==1.9.5.1
PyQt5==5.9.2
PySocks==1.6.7
python-dateutil==2.5.3
python-jwt==2.0.1
pytz==2017.2
pywin32-ctypes==0.1.2
pyzmq==17.1.2
QtAwesome==0.4.4
qtconsole==4.4.0
QtPy==1.4.2
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.7.0
#Shapely==1.6.4.post1
simplegeneric==0.8.1
sip==4.19.8
six==1.11.0
urllib3==1.22
webencodings==0.5.1
wrapt==1.10.11
